🦄 refactor: criando components e reaforando o codigo
